Use system perl or no-op on windows  (#33)
Someone reported on Slack that MakieTeX doesn't work on windows, since Perl_jll doesn't build for the Windows platform. This PR works around that by using system Perl if it exists, or skipping it if it doesn't.

This currently executes on load but it may be worth it to execute on init instead, which would not be a difficult PR but would introduce some complications on how to handle this.